requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
discord.py==1.7.3
idna==3.3
joblib==1.1.0
langdetect==1.0.9
multidict==6.0.2
nltk==3.7
pycodestyle==2.8.0
python-dotenv==0.20.0
regex==2022.4.24
requests==2.27.1
six==1.16.0
textblob==0.17.1
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
